+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - fix bit import -e flag
+- fix - init deletes resolve modules from bit.json
+- throw an error when auto tag components has a newer version
 - fix output when running bit diff after tagging component
 
 ## [13.0.0-dev.10] - 2018-06-08