+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - fix bit import -e flag
 - fix - init deletes resolve modules from bit.json
 - throw an error when auto tag components has a newer version
+- fix output when running bit diff after tagging component
 - fix - after auto-tagging a component with a pending update - it became modified
 
 ## [13.0.0-dev.10] - 2018-06-08